# machinev2/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Law as Code
  version: 0.0.1
servers:
  - url: https://api.server.test/v1
...
paths:
...
  /claims/{bsn}:
    get:
      tags:
        - Claim
      description: Get all claims
      parameters:
        - $ref: "#/components/parameters/pathBSN"
        - $ref: "#/components/parameters/queryOnlyApproved"
        - $ref: "#/components/parameters/queryIncludeRejected"
      operationId: claimListBasedOnBSN
      responses:
        "200":
          $ref: "#/components/responses/ClaimListResponse"
        "400":
          $ref: "#/components/responses/BadRequestErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
  /cases/{bsn}:
    get:
      tags:
        - Case
      description: Get all cases based on a bsn
      parameters:
        - $ref: "#/components/parameters/pathBSN"
      operationId: caseListBasedOnBSN
      responses:
        "200":
          $ref: "#/components/responses/CaseListResponse"
        "400":
          $ref: "#/components/responses/BadRequestErrorResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
  /cases/{service}/{law}:
    get:
      tags:
        - Case
      description: Get all cases based on service and law
      parameters:
        - $ref: "#/components/parameters/pathService"
        - $ref: "#/components/parameters/pathLaw"
      operationId: caseListBasedOnServiceLaw
      responses:
        "200":
          $ref: "#/components/responses/CaseListResponse"
        "400":
          $ref: "#/components/responses/BadRequestErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
components:
  schemas:
    bsn:
      description: Burgerservicenummer of a Dutch citizen
      type: string
      pattern: "^[0-9]{9}$"
      example: "111222333"
    caseID:
      description: Identifier of a case
      type: string
      format: uuid
    claimID:
      description: Identifier of a claim
      type: string
      format: uuid
...
    Error:
      description: The error that occured while processing this request.
      type: object
      properties:
        message:
          type: string
          example: foo went wrong
      required:
        - message
...
    Claim:
      description: Claim
      type: object
      properties:
        id:
          $ref: "#/components/schemas/claimID"
        bsn:
          $ref: "#/components/schemas/bsn"
        service:
          type: string
          description: Specify the service that needs to be executed
        law:
          type: string
          description: Specify the law that needs to be executed
        key:
          type: string
        caseId:
          $ref: "#/components/schemas/caseID"
          nullable: true
        newValue: {}
        reason:
          type: string
        claimant:
          type: string
        oldValue: {}
        evidencePath:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - APPROVED
            - REJECTED
      required:
        - id
        - service
        - key
        - newValue
        - reason
        - claimant
        - law
        - bsn
        - status
...
    ClaimList:
      description: List of all claims
      type: array
      items:
        $ref: "#/components/schemas/Claim"
...
    Case:
      description: Case
      type: object
      required:
        - id
        - bsn
        - service
        - law
        - parameters
        - claimedResult
        - verifiedResult
        - approvedClaimsOnly
        - rulespecId
        - status
      properties:
        id:
          $ref: "#/components/schemas/caseID"
        bsn:
          $ref: "#/components/schemas/bsn"
        service:
          type: string
          description: Specify the service that needs to be executed
        law:
          type: string
          description: Specify the law that needs to be executed
        parameters:
          additionalProperties: true
        claimedResult:
          additionalProperties: true
        verifiedResult:
          additionalProperties: true
        approvedClaimsOnly:
          type: boolean
        rulespecId:
          type: string
          format: uuid
          description: Identifier of the rulespec
        status:
          type: string
          enum:
            - SUBMITTED
            - DECIDED
            - IN_REVIEW
            - OBJECTED
        approved:
          type: boolean
        objectionStatus:
          $ref: "#/components/schemas/CaseObjectionStatus"
        appealStatus:
          additionalProperties: true
    CaseObjectionStatus:
      description: Parameters to set the objection status
      type: object
      properties:
        possible:
          type: boolean
        notPossibleReason:
          type: string
        objectionPeriod:
          type: integer
        decisionPeriod:
          type: integer
        extensionPeriod:
          type: integer
        admissable:
          type: boolean
    CaseList:
      description: List of all cases
      type: array
      items:
        $ref: "#/components/schemas/Case"
...
  parameters:
    pathBSN:
      name: bsn
      in: path
      description: Burgerservicenummer of a Dutch citizen
      required: true
      schema:
        $ref: "#/components/schemas/bsn"
...
    pathService:
      name: service
      in: path
      description: filter for service
      required: true
      schema:
        type: string
        example: TOESLAGEN
    pathLaw:
      name: law
      in: path
      description: filter for law
      required: true
      schema:
        type: string
        example: zorgtoeslagwet
    queryOnlyApproved:
      name: approved
      in: query
      description: If added to URI only approved claims will be returned.
      required: false
      schema:
        type: boolean
        example: true
    queryIncludeRejected:
      name: includeRejected
      in: query
      description: If added to URI rejected claims will be added to the return
      required: false
      schema:
        type: boolean
        example: false
...
  responses:
...
    BadRequestErrorResponse:
      description: Invalid input supplied.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: "#/components/schemas/Error"
                example:
                  - message: foo went wrong
            required:
              - errors
            example:
              errors:
                - message: foo went wrong
    ResourceNotFoundErrorResponse:
      description: Resource not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: "#/components/schemas/Error"
                example:
                  - message: foo went wrong
            example:
              errors:
                - message: foo went wrong
    InternalServerErrorResponse:
      description: Internal server error.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: "#/components/schemas/Error"
                example:
                  - message: foo went wrong
            required:
              - errors
            example:
              errors:
                - message: foo went wrong
...
    ClaimListResponse:
      description: Claim list response
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: "#/components/schemas/ClaimList"
            required:
              - data
...
    CaseListResponse:
      description: Case list response
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: "#/components/schemas/CaseList"
            required:
              - data
...